Update package gen namings and docs
